pyproject.toml
[tool.poetry]
name = "event-logistics-swarm"
version = "1.0.0"
description = "Multi-agent AI system for event logistics automation"
authors = ["Your Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
langgraph = "^0.0.60"
langchain = "^0.1.20"
langchain-openai = "^0.0.8"
langchain-community = "^0.0.38"
openai = "^1.12.0"
chromadb = "^0.4.22"
pgvector = "^0.2.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.1.2"
pandas = "^2.2.0"
numpy = "^1.26.3"
aiosmtplib = "^3.0.1"
email-validator = "^2.1.0"
jinja2 = "^3.1.3"
celery = "^5.3.6"
redis = "^5.0.1"
python-dotenv = "^1.0.1"
httpx = "^0.26.0"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^24.1.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false